pyproject.toml
[tool.poetry]
name = "hillkeeper-bot"
version = "0.1.0"
description = ""
authors = ["CoodingPenguin <coodingpenguin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
discord-py = "^2.6.4"
python-dotenv = "^1.1.1"
redis = {extras = ["hiredis"], version = "^6.4.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
